+++ vertex.cpp
@@ -19,7 +19,7 @@ float Vertex::getY(){
 }
 
 float Vertex::getDistance(Vertex v1, Vertex v2){  // when vertex are board coordenates
-    return (float)sqrt((float)pow( v1.x - v2.x , 2) + (float)pow(v1.y - v2.y, 2));
+    return (float)sqrt(pow( v1.x - v2.x , 2) + pow(v1.y - v2.y, 2));
 }
 
 bool Vertex::Compare(Vertex v1, Vertex v2){ // when vertex are board coordenates
